Massive move data refactor
Still very much in progress, making a commit to have a recovery point.

Logic behind the operation is to massively de-couple starting and
advancing a move from the actions the move will cause the player
character to take. This should allow for more modular code and for
example separately refactor links / special cancels into their own
starter systems instead of dealing with the horror as it is now.